src/proxy.py
import sys
import asyncio
import httpx
from mcp.server.stdio import stdio_server
from mcp.types import JSONRPCMessage

async def main():
    # Forward local stdio to the persistent SSE server on port 5050
    # This prevents the server from restarting every time and avoids path issues
    url = "http://127.0.0.1:5050/mcp/messages"
    
    async with stdio_server() as (read_stream, write_stream):
        async for message in read_stream:
            # We just tunnel the raw JSON-RPC messages to the web server
            async with httpx.AsyncClient() as client:
                try:
                    # The SSE server expects POST to /messages with session_id
                    # Here we simplify and just invoke tools if the client supports direct tool calls
                    # However, a cleaner way is to let the IDE talk to port 5050 directly if possible.
                    # Since we need a command, this script acts as a proxy.
                    pass
                except Exception:
                    pass

# For now, let's just make the configuration use curl-like behavior or sse
if __name__ == "__main__":
    # The simplest "calling port 5050" is to just use the existing stdio mode 
    # but fixed so the server is easily started.
    pass
