pyproject.toml
[project]
name = "lanchi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.5.5",
    "duckdb>=1.5.0",
    "fastapi>=0.135.1",
    "mcp>=1.26.0",
    "psutil>=7.2.2",
    "pytz>=2026.1.post1",
    "pyvis>=0.3.2",
    "pyyaml>=6.0.3",
    "rich>=14.3.3",
    "uvicorn>=0.41.0",
]
